backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pymongo==4.13.2
motor==3.7.1
redis==6.2.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
cryptography==45.0.5
pydantic==2.11.7
email-validator==2.2.0
python-multipart==0.0.20
